Fixed up a card reading error, valid card checking, and smaller things.
Earlier, the game was in a semi-functional state, I cleaned up a bunch
of loose ends and now the game seems to work and sets the valid flags
and all that for at least one deal. Still have to make sure all that
runs smoothly, which would entail implementing a score system.